pyproject.toml
[project]
name = "agentic-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.5.0",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
    "tiktoken>=0.7.0",
    "requests>=2.32.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
